Commit successfull!

+++ BookStoreWebsite/src/com/bookstall/service/BookServices.java
@@ -185,9 +185,15 @@ public class BookServices {
 			request.getRequestDispatcher("message.jsp").forward(request, response);
 			
 		} else {
-			String message = "The book has been deleted successfully.";
-			bookDAO.delete(bookId);			
-			listBooks(message);		
+			if (!book.getReviews().isEmpty()) {
+				String message = "Could not delete the book with ID " + bookId
+						+ " because it has reviews";
+				listBooks(message);
+			} else {
+				String message = "The book has been deleted successfully.";
+				bookDAO.delete(bookId);			
+				listBooks(message);		
+		    }
 		}
 	  }
